update accordion component
